# hazelnet-community/src/main/resources/hazelnet-community-openapi.yml
openapi: 3.0.0
info:
  version: '0.0.1'
  title: 'HazelNet Community'
servers:
  - url: 'localhost:8085'
paths:
...
  /claimlists:
    post:
      summary: Add Claim List
      operationId: addClaimList
      tags:
        - claims
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimList'
      responses:
        '201':
          description: 'Claim List'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimList'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /claimlists/{claimListId}:
    get:
      summary: Get Claim List
      operationId: getClaimList
      tags:
        - claims
      parameters:
        - $ref: '#/components/parameters/ClaimListIdPathParameter'
      responses:
        '200':
          description: 'Claim List'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimList'
  /claimlists/{claimListId}/orders/physical:
    get:
      summary: Get physical orders for this claim list
      operationId: listPhysicalOrders
      tags:
        - claims
      parameters:
        - $ref: '#/components/parameters/ClaimListIdPathParameter'
      responses:
        '200':
          description: 'P'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhysicalOrder'
    post:
      summary: Add a physical order for this claim list
      operationId: addPhysicalOrder
      tags:
        - claims
      parameters:
        - $ref: '#/components/parameters/ClaimListIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhysicalOrder'
      responses:
        '201':
          description: 'Claim List'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalOrder'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
...
  /discord/servers:
    get:
      summary: Get all Discord servers
      operationId: listDiscordServers
      tags:
        - discord
      responses:
        '201':
          description: 'List of Discord servers'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscordServer'

    post:
      summary: Add Discord Server
      operationId: addDiscordServer
      tags:
        - discord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordServer'
      responses:
        '201':
          description: 'Discord Server'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordServer'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
...
  /discord/servers/{guildId}/polls:
    get:
      summary: 'Get polls for this Discord server'
      operationId: listPolls
      tags:
        - discord
        - poll
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '200':
          description: 'All polls for this server'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscordPoll'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a poll on this server
      operationId: addPoll
      tags:
        - discord
        - poll
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordPoll'
      responses:
        '201':
          description: 'Poll'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordPoll'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
...
  /discord/servers/{guildId}/polls/{pollId}/votes/{externalAccountId}:
    put:
      summary: Vote for a specific option
      operationId: setVote
      tags:
        - discord
        - poll
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/PollIdPathParameter'
        - $ref: '#/components/parameters/ExternalAccountPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DiscordPollOptionId'
      responses:
        '200':
          description: 'Vote'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordPollVote'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
...
components:
  parameters:
    GuildIdPathParameter:
      name: guildId
      in: path
      description: The Discord Server's Guild ID
      required: true
      style: simple
      schema:
        $ref: '#/components/schemas/DiscordGuild'
...
    PollIdPathParameter:
      name: pollId
      in: path
      description: The ID of this poll
      required: true
      style: simple
      example: 6121
      schema:
        type: integer
        format: int64
...
    ClaimListIdPathParameter:
      name: claimListId
      in: path
      description: The ID of this claim list
      required: true
      style: simple
      example: 4144
      schema:
        type: integer
        format: int64
...
    ExternalAccountPathParameter:
      name: externalAccountId
      in: path
      description: The ID of the external account
      required: true
      style: simple
      schema:
        type: integer
        format: int64
  schemas:
...
    ClaimList:
      type: object
      required:
        - name
        - displayName
        - claims
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        displayName:
          type: string
          minLength: 1
          maxLength: 256
        name:
          type: string
          pattern: '^[A-Za-z][-A-Za-z0-9]{0,29}$'
        createTime:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T12:21:43Z
        claims:
          type: array
          $ref: '#/components/schemas/ClaimListSnapshotEntry'
    ClaimListSnapshotEntry:
      type: object
      required:
        - stakeAddress
        - claimableProduct
        - claimableCount
      properties:
        stakeAddress:
          $ref: '#/components/schemas/StakeAddressView'
        claimableProduct:
          type: integer
          format: int64
        claimableCount:
          type: integer
          format: int64
    PhysicalOrder:
      type: object
      required:
        - externalAccountId
        - country
        - zipCode
        - city
        - street
        - items
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        externalAccountId:
          type: integer
          format: int64
        country:
          type: string
          example: DEU
          pattern: '^[A-Z]{3}$'
        zipCode:
          type: string
          example: 84323-2031
          minLength: 1
          maxLength: 30
        city:
          type: string
          example: Junkersdorf
          minLength: 1
          maxLength: 200
        street:
          type: string
          example: Burenstrasse 25
          minLength: 1
          maxLength: 500
        items:
          type: array
          minItems: 1
          $ref: '#/components/schemas/PhysicalOrderItem'
    PhysicalOrderItem:
      type: object
      required:
        - productId
        - count
      properties:
        productId:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
...
    DiscordServer:
      type: object
      required:
        - guildId
        - guildName
        - guildOwner
        - guildMemberCount
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        guildId:
          type: integer
          format: int64
          example: 717264144759390238
        guildName:
          type: string
          example: Guild Name
        guildOwner:
          type: integer
          format: int64
          example: 9358923890
        joinTime:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T13:20:57Z
        guildMemberCount:
          type: integer
          format: int32
          example: 150
        ownerAccount:
          type: integer
          format: int64
        tokenPolicies:
          type: array
          items:
            $ref: '#/components/schemas/TokenPolicy'
        stakepools:
          type: array
          items:
            $ref: '#/components/schemas/Stakepool'
        delegatorRoles:
          type: array
          items:
            $ref: '#/components/schemas/DiscordDelegatorRole'
        tokenRoles:
          type: array
          items:
            $ref: '#/components/schemas/DiscordTokenRole'
        settings:
          type: object
          additionalProperties:
            type: string
    TokenPolicy:
      type: object
      required:
        - policyId
        - projectName
      properties:
        policyId:
          $ref: '#/components/schemas/TokenPolicyId'
        projectName:
          type: string
          example: My First CNFT
    Stakepool:
      type: object
      required:
        - poolHash
      properties:
        poolHash:
          $ref: '#/components/schemas/StakepoolHash'
        info:
          $ref: '#/components/schemas/StakepoolInfo'
    StakepoolInfo:
      type: object
      required:
        - hash
        - view
        - ticker
        - name
      properties:
        hash:
          $ref: '#/components/schemas/StakepoolHash'
        view:
          type: string
        ticker:
          readOnly: true
          type: string
        name:
          readOnly: true
          type: string
        website:
          readOnly: true
          type: string
        description:
          readOnly: true
          type: string
    DiscordDelegatorRole:
      type: object
      required:
        - minimumStake
        - roleId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        poolHash:
          $ref: '#/components/schemas/StakepoolHash'
        minimumStake:
          type: integer
          description: Minimum stake in Lovelace
          format: int64
          example: 1000000000
        roleId:
          $ref: '#/components/schemas/DiscordRole'
    DiscordTokenRole:
      type: object
      required:
        - policyId
        - minimumTokenQuantity
        - roleId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        policyId:
          $ref: '#/components/schemas/TokenPolicyId'
        assetFingerprint:
          $ref: '#/components/schemas/AssetFingerprint'
        minimumTokenQuantity:
          type: integer
          format: int64
          example: 1
        roleId:
          $ref: '#/components/schemas/DiscordRole'
    DiscordPoll:
      type: object
      required:
        - creator
        - name
        - displayName
        - description
        - openAfter
        - openUntil
        - options
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        creator:
          type: integer
          format: int64
        channelId:
          $ref: '#/components/schemas/DiscordChannel'
        messageId:
          $ref: '#/components/schemas/DiscordMessage'
        displayName:
          type: string
          minLength: 1
          maxLength: 256
        name:
          type: string
          pattern: '^[A-Za-z][-A-Za-z0-9]{0,29}$'
        description:
          type: string
          minLength: 1
          maxLength: 4096
        createTime:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T12:21:43Z
        openAfter:
          type: string
          format: date-time
          example: 2021-11-10T12:21:43Z
        openUntil:
          type: string
          format: date-time
          example: 2021-11-10T12:21:43Z
        resultsVisible:
          type: boolean
          default: true
        weighted:
          type: boolean
          default: false
        multipleVotes:
          type: boolean
          default: false
        snapshotId:
          type: integer
          format: int64
        requiredRoles:
          type: array
          items:
            $ref: '#/components/schemas/DiscordPollRequiredRole'
        options:
          type: array
          items:
            $ref: '#/components/schemas/DiscordPollOption'
    DiscordPollRequiredRole:
      type: object
      required:
        - roleId
      properties:
        roleId:
          $ref: '#/components/schemas/DiscordRole'
    DiscordPollOptionId:
      type: integer
      readOnly: true
      example: 4929
      format: int64
    DiscordPollOption:
      type: object
      required:
        - text
      properties:
        id:
          $ref: '#/components/schemas/DiscordPollOptionId'
        text:
          type: string
          example: 'Vote for the better one'
        reactionId:
          $ref: '#/components/schemas/DiscordEmoji'
        reactionName:
          type: string
          example: '🔥'
    DiscordPollVote:
      type: object
      required:
        - externalAccountId
        - optionId
      properties:
        externalAccountId:
          type: integer
          format: int64
        optionId:
          $ref: '#/components/schemas/DiscordPollOptionId'
        time:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T13:10:57Z
        weight:
          type: integer
          format: int64
...
    DiscordGuild:
      description: Snowflake ID matching a Discord guild
      type: integer
      format: int64
      example: 717264144759390238
    DiscordEmoji:
      description: Snowflake ID matching a Discord emoji
      type: integer
      format: int64
      example: 817564144253349412
...
    DiscordRole:
      description: Snowflake ID matching a Discord role
      type: integer
      format: int64
      example: 185829035890290
    DiscordChannel:
      description: Snowflake ID matching a Discord channel
      type: integer
      format: int64
      example: 668289032310316
    DiscordMessage:
      description: Snowflake ID matching a Discord message
      type: integer
      format: int64
      example: 468589032310412
    StakeAddressView:
      type: string
      description: The staking address of a wallet, in viewable format
      example: stake1uyhvepgdsd7s07q5cdx9vdemnar4axpe0e9edw2chyd8dpgqnsffq
      pattern: '^stake1[a-zA-Z0-9]{53}$'
    StakepoolHash:
      type: string
      description: The Stakepool hash
      example: be80794a946cf5e578846fc81e3c62ac13f4ab3335e0f5dc046edad4
      pattern: '^[A-Za-z0-9]{56}$'
    TokenPolicyId:
      type: string
      description: The Policy ID
      example: ceb5dedd6cda3f0b4a98919b5d3827e15e324771642b57e0e6aabd57
      pattern: '^[A-Za-z0-9]{56}$'
    AssetFingerprint:
      type: string
      description: The fingerprint of an asset
      example: asset1ls9tt8s03x2gcqjyppx346ct64jts7fnwsmequ
      pattern: '^asset1[A-Za-z-0-9]{38}$'
...
    CreatedResponseHeader:
      type: string
      format: uri
      description: URL of the newly created resource
...